chor(chat): move useToolsMenu hook to libs
