+++ libs/chat-hooks/src/index.ts
@@ -37,4 +37,5 @@ export * from './usePageFileDrag/usePageFileDrag';
 export * from './usePanelMaxWidth/usePanelMaxWidth';
 export * from './useShareLink/useShareLink';
 export * from './useShareRecipientsCount/useShareRecipientsCount';
+export * from './useToolsMenu/useToolsMenu';
 export * from './useViewportWidth/useViewportWidth';
